requirements.txt
pip
numpy
scipy
jpype1
numpy
geopandas
pandas
libpysal
matplotlib
install-jdk